cheqmate_engine/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
pytesseract
Pillow
opencv-python-headless
numpy
scikit-learn
nltk
langchain-text-splitters
joblib
pandas
scipy